Renamet backendApiProxy.ts til proxyMiddlewares.ts.
Flytta qbrick-configen til config-mappa
