+++ Dockerfile
@@ -10,10 +10,10 @@ COPY server/tsconfig.json server/tsconfig.json
 COPY server/package.json server/package.json
 COPY server/yarn.lock server/yarn.lock
 COPY server/server.ts server/server.ts
-COPY server/backendApiProxy.ts server/backendApiProxy.ts
-COPY server/qbrickConfigNoPreload.ts server/qbrickConfigNoPreload.ts
+COPY server/proxyMiddlewares.ts server/proxyMiddlewares.ts
+COPY server/config/qbrickConfigNoPreload.ts server/qbrickConfigNoPreload.ts
 ## TODO: unngå å ha med backendApiProxyMock og testdata på image
-COPY server/backendApiProxyMock.ts server/backendApiProxyMock.ts
+COPY server/proxyMiddlewareMock.ts server/proxyMiddlewareMock.ts
 COPY server/local/testdata.ts server/local/testdata.ts
 COPY server/local/testdata-kurs.ts server/local/testdata-kurs.ts
 COPY server/idporten.ts server/idporten.ts
